# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: NUTTER-XMD WhatsApp Bot API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: pair
    description: WhatsApp pairing operations
  - name: admin
    description: Admin operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /pair/request:
    post:
      operationId: pairRequest
      tags: [pair]
      summary: Request a pairing code
      description: Accepts a phone number, initializes a Baileys session, returns a pair code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PairRequestBody"
      responses:
        "200":
          description: Pair code generated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PairRequestResponse"
        "400":
          description: Invalid phone number
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /pair/qr:
    get:
      operationId: getPairQr
      tags: [pair]
      summary: Get QR code for pairing
      description: Returns the current QR code as a base64 data URL
      responses:
        "200":
          description: QR code data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PairQrResponse"
        "404":
          description: No QR code available yet
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /pair/status:
    get:
      operationId: getPairStatus
      tags: [pair]
      summary: Get pairing status
      description: Returns the current connection/pairing state
      responses:
        "200":
          description: Current status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PairStatusResponse"

  /pair/session:
    get:
      operationId: getPairSession
      tags: [pair]
      summary: Get session ID after successful pairing
      description: Returns the base64-encoded session ID once WhatsApp is connected. Requires the x-pairing-token header returned when pairing started.
      security:
        - pairingToken: []
      responses:
        "200":
          description: Session ID ready
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PairSessionResponse"
        "202":
          description: Pairing still in progress
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Missing or invalid pairing token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /pair/start-qr:
    post:
      operationId: startQrPairing
      tags: [pair]
      summary: Start QR code pairing session
      description: Initializes a Baileys session in QR mode. Returns a pairingToken to use when fetching the SESSION_ID.
      responses:
        "200":
          description: QR session starting
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StartQrResponse"
        "400":
          description: Already connected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /pair/reset:
    post:
      operationId: resetPairing
      tags: [pair]
      summary: Reset pairing session
      description: Clears the current pairing session to start fresh
      responses:
        "200":
          description: Session reset
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /admin/forks:
    get:
      operationId: getAdminForks
      tags: [admin]
      summary: Get all repo forks
      description: Returns list of GitHub users who forked the NUTTER-XMD repo
      security:
        - adminPassword: []
      responses:
        "200":
          description: Fork list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ForksResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /deploy/verify-fork:
    get:
      operationId: verifyFork
      tags: [admin]
      summary: Verify GitHub fork
      description: Checks if a given GitHub username has forked the NUTTER-XMD repo
      parameters:
        - name: username
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Fork verification result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ForkVerifyResponse"
        "400":
          description: Missing username
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  securitySchemes:
    adminPassword:
      type: apiKey
      in: header
      name: x-admin-password
    pairingToken:
      type: apiKey
      in: header
      name: x-pairing-token

  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error

    PairRequestBody:
      type: object
      properties:
        phoneNumber:
          type: string
          description: Phone number in international format (e.g. +254712345678)
      required:
        - phoneNumber

    PairRequestResponse:
      type: object
      properties:
        pairCode:
          type: string
          nullable: true
          description: 8-character pair code. Null initially — poll /pair/status until status is pair_code_ready
        phoneNumber:
          type: string
        pairingToken:
          type: string
          description: One-time token required to retrieve the SESSION_ID from /pair/session
        status:
          type: string
      required:
        - phoneNumber
        - pairingToken
        - status

    StartQrResponse:
      type: object
      properties:
        status:
          type: string
        message:
          type: string
        pairingToken:
          type: string
          description: Token required to retrieve SESSION_ID from /pair/session
      required:
        - status
        - pairingToken

    PairQrResponse:
      type: object
      properties:
        qr:
          type: string
          description: QR code as base64 data URL (data:image/png;base64,...)
        expiresAt:
          type: string
          format: date-time
      required:
        - qr

    PairStatusResponse:
      type: object
      properties:
        status:
          type: string
          enum: [idle, connecting, qr_ready, pair_code_ready, connected, disconnected]
        phoneNumber:
          type: string
          nullable: true
        pairCode:
          type: string
          nullable: true
          description: Pair code once status is pair_code_ready, null otherwise
      required:
        - status

    PairSessionResponse:
      type: object
      properties:
        sessionId:
          type: string
          description: Base64-encoded session credentials to paste into Heroku config
        phoneNumber:
          type: string
      required:
        - sessionId
        - phoneNumber

    ForkVerifyResponse:
      type: object
      properties:
        forked:
          type: boolean
        username:
          type: string
        forkUrl:
          type: string
          nullable: true
        deployUrl:
          type: string
          nullable: true
      required:
        - forked
        - username

    Fork:
      type: object
      properties:
        id:
          type: integer
        login:
          type: string
        avatarUrl:
          type: string
        profileUrl:
          type: string
        forkUrl:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - login
        - avatarUrl
        - profileUrl
        - forkUrl
        - createdAt

    ForksResponse:
      type: object
      properties:
        forks:
          type: array
          items:
            $ref: "#/components/schemas/Fork"
        total:
          type: integer
      required:
        - forks
        - total